Final Code Files/requirements.txt
kafka-python==2.0.2
pandas==2.0.3
numpy==1.24.3
mysql-connector-python==8.0.33
joblib==1.2.0
xgboost==2.0.3
scikit-learn==1.3.0
flask==2.2.5
